requirements.txt
attrs==21.2.0
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
firebase-admin==5.2.0
flake8==4.0.1
Flask==2.0.2
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
httplib2==0.20.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
msgpack==1.0.3
mypy==0.920
mypy-extensions==0.4.3
packaging==21.3
pluggy==1.0.0
proto-plus==1.19.8
protobuf==3.19.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.6
pytest==6.2.5
requests==2.26.0
rsa==4.8
six==1.16.0
toml==0.10.2
tomli==2.0.0
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
